+++ src/time.rs
@@ -1,19 +1,24 @@
 use chrono::offset::{TimeZone, Utc};
-use chrono::DateTime;
+use chrono::{Date, DateTime, NaiveDate};
 use chrono_tz::Tz;
 use chrono_tz::US::Pacific;
 
 /// Return today's date in format 2021-09-16
-pub fn today() -> String {
+pub fn today_string() -> String {
     Utc::now()
         .with_timezone(&Pacific)
         .format("%Y-%m-%d")
         .to_string()
 }
 
+/// Return today's date in Utc
+pub fn today_date() -> Date<Utc> {
+    Utc::now().date()
+}
+
 /// Returns today or date
 pub fn format_date(date: &str) -> String {
-    if *date == today() {
+    if *date == today_string() {
         String::from("Today")
     } else if date.len() == 10 {
         String::from(date)
@@ -23,16 +28,36 @@ pub fn format_date(date: &str) -> String {
 }
 
 /// Parse DateTime
-pub fn datetime_from_str(date: &str) -> DateTime<Tz> {
-    match date.len() {
+pub fn datetime_from_str(str: &str) -> DateTime<Tz> {
+    match str.len() {
         19 => Pacific
-            .datetime_from_str(date, "%Y-%m-%dT%H:%M:%S")
+            .datetime_from_str(str, "%Y-%m-%dT%H:%M:%S")
             .expect("could not parse DateTime"),
         20 => Utc
-            .datetime_from_str(date, "%Y-%m-%dT%H:%M:%SZ")
+            .datetime_from_str(str, "%Y-%m-%dT%H:%M:%SZ")
             .expect("could not parse DateTime")
             .with_timezone(&Pacific),
-        _ => panic!("cannot parse datetime: {}", date),
+        _ => panic!("cannot parse DateTime: {}", str),
+    }
+}
+
+/// Parse Date
+pub fn date_from_str(str: &str) -> Date<Utc> {
+    match str.len() {
+        10 => {
+            let date = NaiveDate::parse_from_str(str, "%Y-%m-%d").expect("could not parse Date");
+            Utc.from_local_date(&date).unwrap()
+        }
+        19 => Utc
+            .datetime_from_str(str, "%Y-%m-%dT%H:%M:%S")
+            .expect("could not parse DateTime")
+            .date(),
+
+        20 => Utc
+            .datetime_from_str(str, "%Y-%m-%dT%H:%M:%SZ")
+            .expect("could not parse DateTime")
+            .date(),
+        _ => panic!("cannot parse NaiveDate: {}", str),
     }
 }